+++ src/common/index.ts
@@ -18,3 +18,4 @@ export * from "./dates.ts";
 export * from "./urls.ts";
 export * from "./html.ts";
 export * from "./aes.ts";
+export * from "./cache.ts";